Beginning Testing
Created a level class meant for each role. There's also some new combat and  UI scripts to start testing out the foundation.

+++ Turn_Based_RPG/Assets/Scripts/Character/Stats/FloatStat.cs
@@ -52,7 +52,7 @@ namespace PsychesBound
 
         }
 
-        public FloatStat(int baseVal)
+        public FloatStat(float baseVal)
         {
             modifiers = new List<IModifier>();
             readonlyModifiers = modifiers.AsReadOnly();
